removes sqlite from internal app state for now creates in-memory player state removes a lot of ui delays needs to save the state to db for resume functionality search results are slooow to navigate

+++ playlist-sync/src/qobuz.rs
@@ -50,6 +50,10 @@ impl<'q> Qobuz<'q> {
             .await
             .expect("failed to refresh config");
         self.client.login().await.expect("failed to login");
+        // self.client
+        //     .test_secrets()
+        //     .await
+        //     .expect("failed to test secrets");
         self.progress.set_message("signed into Qobuz");
     }
 
@@ -136,7 +140,7 @@ impl QobuzPlaylist {
 
     pub fn track_count(&self) -> usize {
         if let Some(tracks) = &self.0.tracks {
-            tracks.items.len() as usize
+            tracks.items.len()
         } else {
             0
         }
